mercado pago

+++ src/main/java/com/tesis/aike/model/entity/UsersEntity.java
@@ -10,7 +10,7 @@ public class UsersEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Id
     @Column(name = "id", nullable = false)
-    private Integer id;
+    private Long id;
     @Basic
     @Column(name = "name", nullable = false, length = 100)
     private String name;
@@ -25,13 +25,13 @@ public class UsersEntity {
     private String password;
     @Basic
     @Column(name = "role_id", nullable = false)
-    private Integer roleId;
+    private Long roleId;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -67,11 +67,11 @@ public class UsersEntity {
         this.password = password;
     }
 
-    public int getRoleId() {
+    public Long getRoleId() {
         return roleId;
     }
 
-    public void setRoleId(int roleId) {
+    public void setRoleId(Long roleId) {
         this.roleId = roleId;
     }
